requirements.txt
fastapi==0.115.2       
uvicorn[standard]==0.31.0
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
structlog==24.4.0
mcp==1.1.0
python-dotenv==1.0.1
prometheus-client==0.21.0
pytest==8.3.3
pytest-asyncio==0.24.0
SQLAlchemy
cerebras-cloud-sdk
httpx
tenacity
docker